Restructure file locations, update imports
